Upgraded hibernate has type result Long instead of Integer for count() operator; fixed class cast exceptions by switching to Long.

+++ web/WEB-INF/classes/src/org/oscarehr/PMmodule/dao/ProviderDao.java
@@ -38,7 +38,7 @@ public class ProviderDao extends HibernateDaoSupport {
 	private Log log = LogFactory.getLog(ProviderDao.class);
 	
 	public boolean providerExists(String providerNo) {
-		boolean exists = (((Integer) getHibernateTemplate().iterate("select count(*) from Provider p where p.ProviderNo = " + providerNo).next()) == 1);
+		boolean exists = (((Long) getHibernateTemplate().iterate("select count(*) from Provider p where p.ProviderNo = " + providerNo).next()) == 1);
 		log.debug("providerExists: " + exists);
 
 		return exists;
